+++ src/error.rs
@@ -29,3 +29,12 @@ pub enum ConfigError {
     #[fail(display = "Field `{}` is missing", path)]
     FieldMissing { path: String },
 }
+
+#[derive(Debug, Fail, PartialEq)]
+pub enum ConfigValidationError {
+    #[fail(display = "Invalid YAML file: {}", err)]
+    InvalidYaml { err: yaml_rust::scanner::ScanError },
+
+    #[fail(display = "`checkers.{}.notifiers` refers to an undeclared notifier `{}`", checker_id, notifier_id)]
+    UnknownNotifier { checker_id: String, notifier_id: String }
+}