+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Better error handling
+* Validate configuration for inconsistencies
+
 #### v0.1.0 - 2018-12-22
 * First MVP release
 * Support of Telegram notifier